requirements.txt
argon2-cffi==21.1.0
asgiref==3.8.1
boto3==1.38.16
botocore==1.38.18
braintree==4.35.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
defusedxml==0.7.1
Django==5.2.9
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-js-asset==3.1.2
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
Markdown==3.8
mutagen==1.47.0
oauthlib==3.2.2
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2021.3
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.12.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.6.1
sqlparse==0.5.4
stripe==12.1.0
typing_extensions==4.14.1
urllib3==2.6.0
whitenoise==6.9.0
